jupyter_fastapi/pyproject.toml
[tool.poetry]
name = "jupyter-fastapi"
version = "0.1.0"
description = ""
authors = ["Mikhail Murtazin <murtazkorpaz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
jupyter-client = "^8.6.2"
jupyter = "^1.0.0"
plotly = "^5.22.0"
numpy = "^1.26.4"
pandas = "^2.2.2"
tensorflow = "^2.16.1"
scipy = "^1.13.1"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
scikit-learn = "^1.5.0"
keras = "^3.3.3"
torch = "2.2.0"
transformers = "^4.41.1"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"
mathematical = "^0.5.1.post1"
kaleido = "0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
